+++ test/test_streaming_summary.ml
@@ -119,7 +119,6 @@ let make_openai_chunk
       ?delta_reasoning
       ?delta_reasoning_details
       ?(delta_tool_calls = [])
-      ?chunk_parse_error
       ()
   : Streaming.openai_chunk
   =
@@ -131,7 +130,6 @@ let make_openai_chunk
   ; delta_tool_calls
   ; finish_reason = None
   ; chunk_usage = None
-  ; chunk_parse_error
   }
 ;;